8.1.5 валидирует хештеги и комментарий
